Cleaned Up Code
Created subroutines so that one routine controls PB logic and
incrementing a value. A limit of the value was created to represent the
limits of our circuits capabilities. The other subroutine then posts the
value to the display. To get the decimal, a colon was used. To actually
get the decimal, we are going to have to do a lot more programming, may
not be worth it to us to do.