ATtiny Program
Decided to use an ATtiny85 to take the output voltage and create a
signal. The signal will pass through an opto. This will allow us to have
isolation between the two sides of the circuit. Right now it takes the
voltage to control a duty cycle, may have it control frequency instead?
Not sure yet.