Created a C-Grade program
Since PCB is not working, created a new file that will be used strictly
for the C-grade circuit. The program has bare minimums to get circuit
to meet teachers expectations. So far, have fixed the PID loop
interrupt issue. However, PID loop is not working, for some reason it
is just taking the calculated Error and setting that as the output duty
cycle.