Most of the way done
At this point, all of the necessary code is in header files that are
being included. Also have changed to using an interrupt fxn to run the
PID loop instead of trying to do it on our own.  As of right now, the
PID loop, initializations, feedback calc, and user input are in their
own header files. Writing to the display had to be done as a subroutine
as a matrix had to be declared. The initializations and writing to
display had to be in the same program as two matrixes could not be
declared.