PID works!
Adam realized that our max output duty we were allowing was too large,
it was not letting the circuit discharge fast enough. So we changed the
max allowable duty cycle from 97% to 78%. Had to pull some terms out of
the subroutines and make them global variables to make it work.